If cell contain has complex structur then `e.target` from the `click` event on the cell can be a child element of `<td>`. If inside of `onSelectRow` the `editRow` will be called the contain of the `<td>` will be overwritten by `<input>` or other control.
In the case the `e.target` will be *modified*. The DOM element `e.target` will be not more the child of the `<td>` and `e.target.parentNode` will be set to `null`.

The old (current) code of jqGrid called `$.jgrid.getCellIndex(td)` *after* the `setSelection`. So inside of [the line](https://github.com/tonytomov/jqGrid/blob/v4.3.1/js/grid.base.js#L37) exception will be fired. In the case not only `c.parentNode.cells` could be not calculated in IE, but even `c` will be undefined because `c` calculated as `c.closest("td,th")[0]` will be `undefined`.

To fix the bug will be suggested to change the order of `onCellSelect` and `setSelection` or `editCell` calls.

Signed-off-by: Dr. Oleg Kiriljuk <oleg.kiriljuk@ok-soft-gmbh.com>