Both ts.p.pager and ts.p.toppager will hold SELECTORS of pagers like "#pager" or "#list_toppager". So we will save the both values *encoded* by $.jgrid.jqID.
The next change is: we will call setPager fuction with id as parameter and not as selector
to be able to generate HTML fragmants based on the id. To get id from selector we will use
$(ts.p.toppager).attr("id"). So the call of setPager fuction will be

setPager($(ts.p.toppager).attr("id"),'_t');

or

setPager($(ts.p.pager).attr("id"),'');

After the changes one can easy to modify the code of setPager to escape ids if needed.

The INPUT value of p.pager can be
     a) the id of the pager (like "pager") - it should be escaped.
     b) the id selector of the pager (like "#pager") - it have to be already escaped.
     c) the DOM or jQuery object which represent the pager. We will get id attribute of it and save as
        ts.p.pager = "#" + $.jgrid.jqID($(ts.p.pager).attr("id"));
     d) wrong parameter - we change it to empty string "".

In the way we can normalize all (or the most) cases of pager usage.

Signed-off-by: Dr. Oleg Kiriljuk <oleg.kiriljuk@ok-soft-gmbh.com>