In locale-en:
Fix the values of `thousandsSeparator`, fix `newformat`, add missing ';' in the body of `S` function.

Add comments which can be helpful by people who translate the texts in other languges.


In locale-de:

Fix the usage of leading zeros in date and the time.
Fix `S` function which practically not used in German.

Add comments which descride the options which will be used.
Add as comments the corresponding format specification which can be used in jQuery UI Datepicker.

Signed-off-by: Dr. Oleg Kiriljuk <oleg.kiriljuk@ok-soft-gmbh.com>