Including the code
    $(this).find(".input-elm").each(function() {
        $(this).triggerHandler("change");
    });

in the `filterData` solve the most problems and allows to implement `closeOnEscape` and `searchOnEnter` options.

New option `uiButtons` (with additional `uiButtonOptions` option) makes jQuery UI style of all buttons.

New `texts` option allows to change the default '+ {}', '+' and '-' texts of buttons and to define lokalized tooltips.
Probebly one can add additional options to define tooltips on "Query", "Find" and "Reset" buttons.

The implementation of `reDraw` was extended to improve keyboard support and hold the focus.

Signed-off-by: Dr. Oleg Kiriljuk <oleg.kiriljuk@ok-soft-gmbh.com>