fix for issue #131
When grouping is enabled, jqGrid adds the grouping fields to the sort
columns. However convention says that sort column should contain the index
values if defined in the colModel.

The fix checks whether an index is defined in the respective colModel
definition and uses that if present.

This works on my machine!