tabindex attribute changed from 1 to 0.
Having the tabindex set to 1 was causing the focus to jump around
the screen quite a bit e.g. hitting <tab> would go straight to the
jqGrid, but subsequent <tab> key presses would then jump back up
to the main navigational links.

Changing the tabindex to zero meant that jqGrid <table> can still
receive the focus, but receives it according to the normal document
flow.