Fixing wrong encodding of rowid and optimizing the code
The current version of `$.fn.fmatter.rowactions` calls other jqGrid
methods like `editRow` with encoded rowid (unneeded call of `$.jgrid.jqID`).
The suggested code fix the problem.

Additionally the correct code place in the cell `onclick` attribute with
fixed values of `rowid` and `pos`. So changing of column order (with respect
of `columnChooser`) or changing of `rowid` (`.attr("id",newRowId)`) follows to
wrong `onclick` code.

I suggest to use `this` in the code of `onclick` and get all needed
information from `this`.