Solves issues with IOS 7
Solves an issue with IOS 7 which seems to not work very well with the
namedItem function of the HTMLCollection.
It causes errors while saving rows through the inline edit mode per
example.
A new function was added into grid.custom.js called getNamedItem to wrap
the calls to namedItem() and replacing it for item().