fixing unneeded usage of formatter if formatDisplayField are defined
If `formatDisplayField` are defined than it could returns the string which is incorrect for `formatter`. For example if one want group date by month one want to display in grouping row soththing like `Oct-2013` or `2013-10`, but the current implementation of `groupingRender` try to call date `formatter` with the value returned by `formatDisplayField` callback. So the date are displayed with `Nan` or `Undefined`. To fix the problem I suggest to call `formatter` only in case when `formatDisplayField` callback in not defined.