Display only the next grouping level or the data in case of expanding of groping header
The current implementation of `groupingToggle` shows **all** sub-groping headers and the data itself on clicking on the "+" icon in the groping header. Such behavior seems me not practical in case of usage deep multilevel groping.

The suggested changes the behavior so that only the headers of the next groping sub-level will be displayed in expanding of group. Only if no sub-group exist then the data will be displayed.