fix processing of click event on disabled pager button
The processing of click event on disabled pager button (at least in IE) follows `selrow` parameter will be set to `null` (inside of `clearVals`) without removing `"ui-state-highlight"` class from selected row (or selected rows in case of usage `multiselect: true`). See [the answer](http://stackoverflow.com/a/20193776/315935) or [the post](http://www.trirand.com/blog/?page_id=393/bugs/more-as-one-rows-could-be-selected-with-multiselect-false-if-disabled-pager-button-will-be-clicked-before/#p29883) for more details.