fix applying searching without saving the data in rule
It's not really important whether the control have jQuery UI Autocomplete or not. It's important just that **at the moment when the user click "Search" button the current data from the input fields should be used**. One need just trigger `change` event on the input which have the focus to be sure that `rule` will contain the data which the user see in the Search dialog. It's important that the `change` event will be immediately processed and the `rule` be updated. The state of jQuery UI Autocomplete at the moment is not important.