Fix bug where the click event is bound multiple times.
When using virtual scrolling and subgrids, when scrolling down and the
grid is adding new rows, i starts over at 1 and rebinds the click event
for the expand button. This causes the click event to get fired multiple
times (as many times as additional rows were added when scrolling).
Adding a simple unbind before binding will eliminate this, and allow the
expand button to work properly. The "click" event was namespaced to
eliminate future bugs if namespacing was used.