Separate custom sort function from sorttype to fix filtering by dates.
If a custom sorttype is used to perform multi column sorting and the column is a date the filtering does not handle the date appropriately.  By using the formatter='date' to identify a date column with a custom sorttype we are able to properly filter a date using the display format used while still having the ability to sort by multiple columns. (ie. Date + SeqenceNo )