Toolbar filters not always ready when triggered.
it has been found when binding select lists to a service method that they are not always ready when the toolbar is triggered and the default value was being ignored and the data was not filtered correctly.

we have taken the chance to trigger the toolbar and flag it as a initial trigger so when the toolbar select elements are not in place do to how async http calls work. that we are allowed to fall back on the default values.