worked out another performance enhancer
the save row method was looping through a grid with ~45 columns. the first 10 at most have a chance of being editable. using a control list we can control the looping for only the editable fields and once there is no more fields in the list we can stop looping and increase performance.