call onCellSelect and editCell independent on beforeSelectRow
The changes are described [here](http://www.trirand.com/blog/?page_id=393/bugs/calling-callback-after-jquery-event-which-return-a-value#p31346). The changes makes the results returned from `beforeSelectRow` and `jqGridBeforeSelectRow` be influence *only on selection* of the row. `onCellSelect` and `jqGridCellSelect` and `editCell` (the last only in case of usage `cellEdit: true`) be called always.