implement new autoResizable: true property of colModel and the method autoResizeColumn
The property autoResizable: true defined in a column in colModel makes
formatter and unformatters works in a little another way. The cell
content and the content of the column header will be wrapped in <span>.
It allows to calculate the width of the text and so the optimal width of
the column. On double-click on the column resizer the optimal column
width will be set. Additional grid options allows to control the
calculation process. The most important option is boolean
autoResizableCompact.