Updated the Excel Export with a means to serialize additional post data
to use this enhanced code pass a function for the data option. First parameter data is an object. Second parameter is a boolean and I use this as a variable called 'get'. this way if my serialize method is converting the data to json for a regular post. if true we leave it as an object