implement new labelAlign and labelClasses properties of colModel
The new `labelClasses` property of `colModel` can be used like existing
`classes`  property, but it will apply the class to the label (the
column header).
The new `labelAlign` property `colModel` can be used like existing
`align` property, but it will specify alignment of the label (the column
header). It can have the value `"likeData"` additionally to `"left"`,
`"center"` and `"right"` allowed for the `allign` property. The property
`labelAlign: "likeData"` means the usage of `allign` property to align
the label (if no `allign` property are defined for the column then the
default `"left"` alignment will be used for the data and for the label).