deselect on click on selected row in single-row select mode
new option `singleSelectClickMode` allows to control deselection on
clicking on the previously selected row. The default value is
`"toggle"`, which follows toggle selection/unselection on click on the
row. The have old behavior one need to specify any other value for
`singleSelectClickMode` option, for example `singleSelectClickMode:
"selectonly"` or `singleSelectClickMode: ""`. It's important to stress,
that the behavior of `setSelection` method is **not changed**. Multiple
calls of `setSelection` will not deselect the row.