Additional small improvements which reduced the size of jquery.jqGrid.min.js
Now `jquery.jqGrid.min.js` have 274.124 bytes which is less then 274.310
bytes of jqGrid 4.6.0 and 281.888 bytes of jqGrid 4.7.0, but the code
have more functionality and in spite of `jquery.jqGrid.js` is 18.478
byes larger as `jquery.jqGrid.js` of jqGrid 4.7.0. One should still
remark that Compressed size (gzipped) of of jquery.jqGrid.min.js is
still higher as of gzipped jquery.jqGrid.min.js of jqGrid 4.6.0 or
jqGrid 4.7.0.