Improvements in pager options of jqGrid, navGrid and inlineGrid
1) one can use `pager: true` option of jqGrid. In the case new `<div>`
with unique `id` will be generated, placed on the `<body>` and the
`pager` option will be modified to id selector of the new `<div>`
2) one allow to use `navGrid` skipping of `pager` parameter. `navGrid`
will will create the navigator bar on all pagers of the grid (on one or
two pagers depend on the values of `pager` and `toppager` option of
jqGrid).
3) one allow to call `inlineNav` skipping of `pager` parameter.
`inlineNav` will will create the navigator bar on all pagers of the
grid.
4) one allows to call `inlineNav` without previous calling of `navGrid`
to create the empty pager.  `inlineNav` will call `navGrid` itself if
it's required.