more bug fixes, make removemodal:true style default for form editing
some additional fixes need be made in the form editing to remove any
usage of *outer variables* inside of 'click' and other callback
handlers. It will make DOM of grid not locked. Additionally it will
solve the problem with usage of options **from the first call** of the
methods when the dialog will be created.