support of editable property of colModel as function
The callback function can return `true` which allows editing in inline
and cell editing mode. Form editing supports some additional string
values: `"hidden"`, `"disabled"` and `"readonly"`. All three value
follows to creating row in the form, the value of the column will be
sent to the server. The value `"disabled"` can be used with many types
of data. It set the `disabled` and `readonly` property on the control
additionally to adding `"ui-state-disabled"` class to data part of the
row and to the label part. The value `"readonly"` can be practical for
text input controls.