change options of onPaging callback, new event jqGridPaging
`onPaging` and jqGridPaging get additional `options` parameter which is
object with 5 properties `newPage`, `currentPage`, `lastPage`,
`currentRowNum` and `newRowNum`.

The value of the first parameter corresponds to old documentation of
`onPaging`: the first parameter is the string with one from the
following vales: `"first"`, `"last"`, `"prev"`, `"next"`, `"records"` or
`"user"`. The version 4.7 used in reality the `id` of the pager buttons
instead of `"first"`, `"last"`, `"prev"`, `"next"`. So the strings
`"first"`, `"last"`, `"prev"`, `"next"` could be appended with `"_"` and
the id of the pager (or toppager).