fixes and changes to support editing of TreeGrid
After moving the most parts of `setTreeNode` to base and fmatter modules
the performance of TreeGrid is better. The changes required more
changing in the code of inline and form editing. Moreover the code of
cell editing is now adjusted too to allows full editing of tree grid
like it was allows already for inline and form editing. The changes
should work. I will make some additional optimizations later to reduce
existing code duplicates in different editing modules.