Bug fix in grouping
addLocalData should always reset groupingView.groups for local data. The
reason: the data will be displayed based on the order of groups in
groupingView.groups array. addLocalData do sorting of data taking in
consideration of groupField and groupOrder which could changes the order
of items. If the grid will be created using `datatype: "jsonstring"`
then the initial order of data could not corresponds the parameters of
groupingView.  The old code creates groupingView.groups once on loading
and don't changes it till `search` or `resetsearch` will be set to true.
It's not correct. The changes recreate groupingView.groups information
every time (after sorting of data).