performance improvements: new option quickEmpty, new interface getCellBuilder of formatters
The formatting of columns which predefined formatters ("number",
"integer", "date", "checkbox", "checkboxFontAwesome4", `"showlink"`,
`"select"`) is relatively slow in version 4.8. The main reason is the
usage of `getGridRes` **for building of every cell** instead of getting
the information once for column during filling of multiple rows in the
grid. The method `getGridRes` are used to get language specific
formatters. New `getCellBuilder` interface is added to formatters. It
allows to generate *closure* which will be used during filling of
multiple rows of the grid. It improves performance of the filling
essentially.

`emptyRows` can take much more time as new way of deleting of rows. The
performance is essentially slow in case of usage jQuery UI js (not CSS
only) in specific versions. New option `quickEmpty: true` is added in
free jqGrid which speed up deleting of rows. The local paging, sorting
and `reloadGrid` in general works now much quickly. **The option will
don't work if `deepempty: true` set.** If one will have some problems
(some kind of memory leaks) with the usage of new style of deleing then
one can add `quickEmpty: false` to switch off the new style of deleting
and to use jQuery.remove as like it was before.