large and important changes, XML/JSON input parsing are rewritten
new internal parameters: arrayReader, iColByName, iPropByName,
reservedColumnNames, rowIndexes. Internal methods buildAddPropMap,
buildArrayReader

replacing addXmlData and addJSONData to one common function readInput.
Part of the code are moved in new $.jgrid.parseDataToHtml which will be
used in addRowData.

normalizeData and addRowData need be more changed. One will remove all
from normalizeData and probably use some more common part in
normalizeData, readInput and addRowData in separate new helper function.

usage new internal getTdFromTarget in click, dblclick and contextmenu
handler. the code will be more safe. getCellIndex will be not used
internally in the future.

Code optimizations in autoResizeAllColumns to improve the performance.

Performance improvement in getGridRowById and getInd.

The changes started in the post will be continued. The current code need
be more tested of case, but all me tests looks good now.

The "cb" column with muliselect like subgrid or rownumber could be soon
have any position in the grid. For example be on the right side of the
grid. The "cb" column will be optional.