new multiselectPosition option: "left", "right", "none"
The new multiselectPosition option can be used in combination with
`multiselect: true`. It can have the following values:
- "left" (default - the old style).
- "right" - it places the column "cb" with checkboxes for multiselect on
the right side of the grid
- "none" or any other value remove checkboxes. but allow still use
multiselection. The option makes selection quickly. If one have large
grid without paging then the option is recommended.

The internal code of free jqGrid is rewritten in many places. Now the
columns `"rn"` (`rownumbers:true`), `"cb"` (`multiselect:true`) and
`"subgrid"` (`subGrid:true`). So it's possible to reorder the columns
using `remapColumnsByName` or the old `remapColumns` to reorder the
position of "cb" column dynamically and jqGrid should still work.

For example if the column "userName" is the first column in `colModel`
and one uses `multiselect: true` option then one can use
`$("#grid").jqGrid("remapColumnsByName", ["userName", "cb"], true)` to
place multiselect column AFTER the column "userName". In the way `cb`
column can be somewhere in the middle of the grid and jqGrid will still
work.