allow keyboard events inside of modal for controls with absolute position
Allow keyboard events inside of modal for controls which use add itself
under the body as parent and which uses position:absolute to be inside
of modal dialog. An example are select2, autocomplete, different menus
and so on. See the issue https://github.com/free-jqgrid/jqGrid/issues/59
as an example