column resizer is changed: color and max position
The resizer of columns can be moved on the right of the right edge of
the grid.

The color of resizer will be get from the border color of
guiStyles..resizer which is ui-widget-header for jQueryUI
(guiStyles.jQueryUI.resizer).

The constant 33 used as the minimal width of columns during resizing are
replaced to new option minResizingWidth with default value 10.