improve results in case of multiple calls of setGroupHeaders
The fix makes headers look good in case of multiple calls of
`setGroupHeaders`. jqPivot use such calls in case of more as two
yDimension or in case of multiple aggregates. I'm not sure that the
results will be always correct, but the suggested code works not less
correct as the previous one and it produce pretty good results.