fix of visibility of multiselect checkbox in column header
see https://github.com/free-jqgrid/jqGrid/issues/70

The problem exist in Firefox in case of usage non 100%, 200% and so on
zoom. The fix remove the standard padding 2px from the column header of
the multiselect checkbox. Because of usage `text-align: center;` applied
on the header such change not changes the position of the checkbox, but
it seems to solve the described problem in Firefox.