More changes in pivotSetup method
The code now is much more readable. It's still preliminary version. Some
options not will be applied. Grouping by group (by usage `totals: true`
in `yDimension` or by specifying `totalHeader` and `totalText`) build
column headers, but the aggregation values will be not yet calculated in
the grid and be displayed as 0 or stay empty.

Pivot work only with sorted data (at least by Y) to be able to group the
data. So one should include `sorttype` property in items of
`yDimension`. One should use typically `sorttype: "number"`, `sorttype:
"integer"`. Skipped value of `sorttype` means sorting as strings.

The main reason of the last reorganization of the code is making it more
readable and more easy testable.