working new pivot
The code seems to work now with all features. I need only add support of
`converter` callback in `yDimension`, but it will be probably not needed
in the most cases. The names of columns will be build NOT based on the
texts from the input data. So the data can be just saved like it be
without removing spaces for example.