filling pivotInfos for totals
1) `pivotInfos` is extended with `colType` property which have 0 value
for the standard columns, 1 for group totals and 2 for row totals
2) filling `pivotInfos` for group totals and row totals. Because one can
have more as one y items in the case, the `pivotInfos` don't have `y`
and `iY` in the case. Instead of that there are properties `ys` and
`iYs`, which are arrays with `y` values and `iY` values used in the
totals.

One should be carefully with processing of `pivotInfos` in `cellattr`
and custom formatters. The property `pivotInfos` will be set only of
columns which have some values. If no data are set for the column then
the  `pivotInfos` is also undefined.