handle hDiv/fhDiv scrolling on keyboard navigation
The hdivs can be scrolled if the filter toolbar exist. If one uses Tab
key for jumping the focus between input fields of the filter toolbar
then the hDiv/fhDiv could be scrolled even the divs have no scrollbars.
By the usage of `scroll` event handler we prevent scrolling of fhDiv and
to synchronize bDiv scroll position with the scroll position of hDiv.
One can improve the keyboard navigation inside of filter toolbar with
frozen columns later to have better horizontal scrolling and the order
of fields. The current commit fixes only clear bug in the existing code.