apply labelClasses on th instead of div, $.jgrid.isCellClassHidden
The main problem is the following: the `classes` or `labelClasses` can
specify classes which **hide** the column. It could be
`"ui-helper-hidden"` or bootstrap classes like `"hidden-xs"` which helps
to implement responsive grid which hide some less important columns on
case of low resolution or low width of grid. The old implementation of
`setGridWidth` and `setInitialColWidth` methods used only the value of
`hidden` property. Moreover the classes from the properties  `classes`
or `labelClasses` was not applied on the first row of the grid body and
on another columns of the column header (searching toolbar, grouping
columns). The fix solves the problem.