instantiate parameters of ajax for dataUrl
The old code could work wrong it there are more as one column with
`searchoptions.dataUrl`. Multiple Ajax request will be made, but the old
`success` closure function used just `cm`, `ci`,  `stbl`, `soptions`
variables from the outer scope. As the result the values could be from
the *last* column which have `searchoptions.dataUrl`. The most simple
way to *instantiate* the values of the variables for every ajax call is
the usage of `context` parameter of `$.ajax` exactly what one do in
`$.jgrid.createEl` method.