fix problem with missing common icon class for minus
The problem can exist if one specify treeIcons.minus icon as parameter
of jqGid. In the way p.treeIcons.commonIconClass will be removed by
removeClass because it's the part of standard value of p.treeIcons.plus,
but it will be not set by addClass.

Remark: The existing code need be optimized.