add options to getRowData
the second parameter (or the first if it has the type object) can be
object with options. Two boolean options are supported currently:
`skipHidden` and `includeId`. If one use `{includeId: true}` then
`prmNames.id` specifies the name of `id` property. If some column of
`colModel` have `key: true` and it will be exported then no additional
`id` will be included.