improve compatibility with jQuery 3.0.0 alpha1
jQuery 3.0.0 alpha1 seems to have bug in creating of ´<th>` and `<tr>`
elements. jQuery use another code if the text of the code is just
`"<th></th>"` or `"<tr></tr>"`. Independend from the clear
interpretation of the existing problems as a bug of jQuery alpha I
decided to make the changes.