performance improvements of parseDate
the usage of `$.extend(true, {}, ...)` is much slowly as the usahe of
`$.extend({}, ...)`. The deep version is not really required in almost
all cases. So we stay by the short version. We can consider to make more
optimizations later to get scalars like "formatter.date.parseRe" or
"formatter.date.srcformat" instead of getting OBJECT "formatter.date"
and usage properties of the object later.