add support of subGridOptions.hasSubgrid callback
new callback could be practical if the input data for the row contain
information whether the row contains subgrid data or not. The only
parameter of `hasSubgrid` is the `options` parameter with properties
`rowid`, `iRow`, `iCol`, `data`. The `data` parameter contains
previously read data for  the row. If  `hasSubgrid` returns `false` then
no "+" icon for the expanding subgrid will be added in the "subgrid"
column.