improvement of the performance of frozen columns
The performance of expanding/collapsing of data grouping, Inline
editing, cell editing in essensually improved in case of usage frozen
columns. The height of individual rows will be recalulated now only if
it's required.

The event `jqGridResetFrozenHeights` has now optional parameter which
allows to specify posttions and height oof which divs (hDiv, bDiv, sDiv)
should be recalculated. Additionally one can specify the height of which
rows of hDiv/bDiv need be recalculated. The parameter of
`jqGridResetFrozenHeights`  is object with 3 properties `header`,
`resizeFooter` and `body`. The property `resizeFooter` is boolean. It
specify whether the position and height of the div need be recalculated.
The format of `header` and `body` property is identical. It's object
with two properties: boolean `resizeDiv`, which have the same meaning
like `resizeFooter` and the property `resizedRows`, which is array from
two integer elements. It allows to specify interval of rows which
heights need be recalculated. The first element of `resizedRows` array
can be -1, which mean no recalculating of hight of individual rows. 0 or
positive value of `resizedRows[0]` means the index (`rowIndex` of
`<tr>`) of the first row, which height need be recalculated. The second
element of `resizedRows` array can be -1, which mean that the height of
all element starting with `resizedRows[0]` need be recalculated.
Non-negative value of `resizedRows[1]` means the index (`rowIndex` of
`<tr>`) of the last row, which height need be recalculated.

For example we can make some changes in top-pager, like hiding the left
(naviator part) of the top-pager. As the result one would need adjust
position of all divs, but the adjustment of the height of individual
rows are not required. Thus one can use the following code

```JavaScript
$($grid.jqGrid("getGridParam", "pager") + "_left," +
$grid.jqGrid("getGridParam", "toppager") + "_left").hide();
// after changing the pager the positions of the frozen divs could be
changed, but one don't need to recalculate
// the height of every row inside of the divs
$grid.triggerHandler("jqGridResetFrozenHeights", [{
header: {
resizeDiv: true,      // recalculate the position and the height of hDiv
resizedRows: [-1, -1] // don't recalculate heights of every row inside
of hDiv
},
resizeFooter: true,       // recalculate the position and the height of
sDiv
body: {
resizeDiv: true,      // recalculate the position and the height of bDiv
resizedRows: [-1, -1] // don't recalculate heights of every row inside
of bDiv
}
}]);
```