new hasAllClasses, new maxHeight param, width fixes (scrollbar)
add new `$.jgrid.hasAllClasses` method for usage in `groupingToggle`.

new option maxHeight allows to set CSS property `max-height` on bDiv.
It's very practical setting in combination with `height: "auto"`
(default value) and probably large number of rows. In case of small
number of rows one will have the same results as with `height: "auto"`
only, but one cen be sure that that the height will be not higher as
maxHeight for the large number of rows. See
http://stackoverflow.com/a/5896432/315935.

implement some fixes in setting of width depend on existing vertical
scrollbar after inserting data in the grid wich have no `width`
parameter (or `width: "auto"` or `width: "100%"`)

change paremters used for triggering of `jqGridResetFrozenHeights` event
in `setGridWidth`.