small fix in setGroupHeaders, setFrozenColumns have essentially improved performance
`setGroupHeaders` is fixed to have `<thable>` always before `<btable>`.

The code of setFrozenColumns is essentially improved to have better
performance. Setting of height on rows inside of internal `fixDiv`
function is the bottleneck. It set now height only on **visible** rows.
It improves essential the performanmce. Additionally the parameters of
`jqGridResetFrozenHeights` are changed to allow to specify whether
recalculation of rows are required or not.

The main problam is the following. One uses typically paging of data to
display large number of rows, but there are two important exceptions:
TreeGrid and data grouping (in collaped form: groupCollapse: true). One
can have many hidden rows in the grid, but to display only small subset
of rows. Exactly in the case the new changes in `setFrozenColumns`
improve the performance of frozen columns.