add `u1000` date format
The format "u" and "U" parse the input data like `1418297439` as
`12/11/2014 11:30:39` (displayed in `newformat: "m/d/Y H:i:s"`). The new
format `u1000` can be halpful to parse input values like
`1418297439000`, which value is the time in Milliseconds since the Unix
Epoch (January 1 1970 00:00:00 GMT). See
http://stackoverflow.com/a/33652984/315935 for the problem existing in
Safary on MAC where the trick with the usage of format `srcformat:
"U/1000"` (or `srcformat: "U/"` or `srcformat: "U "`) not work.