Add support of editable:"hidden" for inline editing
Thanks @singe3 (http://stackoverflow.com/users/3751590/singe3) for the
suggestion. See
http://stackoverflow.com/questions/6055712/posting-non-editable-values-in-jqgrid/6057523?noredirect=1#comment55164298_6057523.

One can set `editable:"hidden"` in any hidden or visible column to force
sending of the content of the column to the server on row editing. Old
versions of jqGrid allows only `hidden: true, editable: true, editrules:
{ edithidden: false }` settings to send the content of **hidden** column
to the server. The previous version of free jqGrid support sending of
columns having `editable:"hidden"` (see [the
wiki](https://github.com/free-jqgrid/jqGrid/wiki/editable-property-of-colModel-as-function)),
but the feature was supported till now onle in *form editing* and not in
inline editing mode.

Later I add sending `additionalProperties` to the server if the property
have the same `editable:"hidden"` property.