Change escaping of edited data, fix width of grid
1) `autoencode:true` not follow encoding of edited data by
`$.jgrid.htmlEncode`. It can be bad if one send data via JSON encodding
for example.

2) jqGrid have now new option `autoEncodeOnEdit` (default value `false`)
which can be used to force old behavior. Even if `autoEncodeOnEdit:true`
is used, there are minor differences. The encodding will be done now
directly before sending the data. The validation functions and other
callbacks will see not-encodded data.

3) The new functions `$.jgrid.oldEncodePostedData` and
`$.jgrid.oldDecodePostedData` be used now for encodding the data which
need be posted on editing. The functions `$.jgrid.htmlDecode` and
`$.jgrid.htmlEncode` encode now the characters `'` and `/`.

4) main grid table (`.ui-jqgrid-btable`) and header hable
(`.ui-jqgrid-htable`) have now `width:1px` CSS property, which force
usage of `table-layout: fixed` and fix small problem with differences in
the width of the grid and the table with column header. The problems
were in Google Chrome, in case of zoom factor not equal 100% (or 200%,
...). See [the answer](http://stackoverflow.com/a/33875600/315935) for
more details.

5) Small fixes in `title` which will be set inside of `setRowData`
method.

Thank to [Andrus](http://stackoverflow.com/users/742402/andrus)
([kobruleht](https://github.com/kobruleht)) for helpful discussion about
the above changes, preparing the demos and testing preliminary versions
of suggested changes.