Fix wrong calculation of scrollOffset in IE
If `bDiv.clientWidth === 0 && bDiv.clientHeight === 0` then
`bDiv.offsetWidth` could be set to the width of bDiv, but
`bDiv.clientWidth` will be 0. The problem exist in IE and could follow
to wrong calculation of `scrollOffset` inside of
`$.jgrid.fixScrollOffsetAndhBoxPadding`. Setting of `min-height` prevent
the problem.

Thanks @Yagg (https://github.com/Yagg) for reporting of the problem in
https://github.com/free-jqgrid/jqGrid/issues/182