Fix in $.fn.fmatter.rowactions, to avoid copying sub-objects of cm.formatoptions by reference
This fixes a bug where if you add rows using $.jgrid.addRow(), and you have a column
with 'actions' formatter, and you click the edit button in the actions column, the
'oper' param of the AJAX request for saving the edit will still be 'add'. This happened
because cm.formatoptions.extraparams passed by reference into 'op', and then oper=add
was set to it. Now it is no longer passed by reference, due to using deep clone first.