Fix: colMenu sort was not working with multisort
Using the column menu sorting with multiSort: true on the grid resulted
in a toggle from asc to desc to not ordered instead of following what
the user clicked on. I have modified the multiSort function to include a
"sor" parameter that will dictate the sort order if it is provided,
instead of causing the regular toggle.

Included in this commit is also an additional french language fix that
was pointed out by one of my coworkers.